Python/Projects/database/requirements.txt

django
python-slugify
psycopg2